feat: Allow using openAI routes without partition
We now allow to directly request the LLM through the openAI endpoints.
This is useful when we want to interact with the LLM without relying on
documents, e.g. for translation, rephrasing, etc.
To use it, the client must simply not specify the `model`.

Note that is this approach, no system prompt will be used, it is
entirely up to the client.